requirements.txt
requests~=2.32.5
beautifulsoup4~=4.12.3
openai~=2.21.0
python-dotenv~=0.21.0